+++ crates/fuel-gas-price-algorithm/src/v1/tests/algorithm_v1_tests.rs
@@ -8,13 +8,13 @@ fn calculate__even_profit_maintains_price() {
     let starting_cost = 500;
     let latest_gas_per_byte = 10;
     let da_gas_price_denominator = 1;
-    let block_bytes = 500;
+    let block_bytes = 500u64;
     let starting_reward = starting_cost + block_bytes * latest_gas_per_byte;
     let updater = UpdaterBuilder::new()
         .with_starting_exec_gas_price(starting_exec_gas_price)
         .with_starting_da_gas_price(starting_da_gas_price)
         .with_da_p_component(da_gas_price_denominator)
-        .with_total_rewards(starting_reward)
+        .with_total_rewards(starting_reward as u128)
         .with_known_total_cost(starting_cost as u128)
         .with_projected_total_cost(starting_cost as u128)
         .with_da_cost_per_byte(latest_gas_per_byte as u128)
@@ -38,7 +38,7 @@ fn calculate__negative_profit_increase_gas_price() {
     let latest_gas_per_byte = 10;
     let da_p_component = 100;
     let da_d_component = 10;
-    let block_bytes = 500;
+    let block_bytes = 500u64;
     let last_profit = -100;
     let last_last_profit = 0;
     let arb_value = 1000;
@@ -49,7 +49,7 @@ fn calculate__negative_profit_increase_gas_price() {
         .with_starting_da_gas_price(last_da_gas_price)
         .with_da_p_component(da_p_component)
         .with_da_d_component(da_d_component)
-        .with_total_rewards(smaller_starting_reward)
+        .with_total_rewards(smaller_starting_reward as u128)
         .with_known_total_cost(starting_cost as u128)
         .with_projected_total_cost(starting_cost as u128)
         .with_da_cost_per_byte(latest_gas_per_byte as u128)
@@ -61,11 +61,12 @@ fn calculate__negative_profit_increase_gas_price() {
     let actual = algo.calculate(block_bytes);
 
     // then
-    let da_p_comp = last_profit / da_p_component;
+    let da_p_comp = last_profit / da_p_component as i128;
     let slope = last_profit - last_last_profit;
-    let da_d_comp = slope / da_d_component;
-    let expected =
-        starting_exec_gas_price as i64 + last_da_gas_price as i64 - da_p_comp - da_d_comp;
+    let da_d_comp = slope / da_d_component as i128;
+    let expected = starting_exec_gas_price as i128 + last_da_gas_price as i128
+        - da_p_comp
+        - da_d_comp;
     assert_eq!(expected as u64, actual);
 }
 
@@ -78,7 +79,7 @@ fn calculate__positive_profit_decrease_gas_price() {
     let latest_gas_per_byte = 10;
     let da_p_component = 100;
     let da_d_component = 10;
-    let block_bytes = 500;
+    let block_bytes = 500u64;
     let last_profit = 100;
     let last_last_profit = 0;
     let arb_value = 1000;
@@ -89,7 +90,7 @@ fn calculate__positive_profit_decrease_gas_price() {
         .with_da_p_component(da_p_component)
         .with_da_d_component(da_d_component)
         .with_starting_da_gas_price(last_da_gas_price)
-        .with_total_rewards(larger_starting_reward)
+        .with_total_rewards(larger_starting_reward as u128)
         .with_known_total_cost(starting_cost as u128)
         .with_projected_total_cost(starting_cost as u128)
         .with_da_cost_per_byte(latest_gas_per_byte as u128)
@@ -103,11 +104,12 @@ fn calculate__positive_profit_decrease_gas_price() {
     let actual = algo.calculate(block_bytes);
 
     // then
-    let da_p_comp = last_profit / da_p_component;
+    let da_p_comp = last_profit / da_p_component as i128;
     let slope = last_profit - last_last_profit;
-    let da_d_comp = slope / da_d_component;
-    let expected =
-        starting_exec_gas_price as i64 + last_da_gas_price as i64 - da_p_comp - da_d_comp;
+    let da_d_comp = slope / da_d_component as i128;
+    let expected = starting_exec_gas_price as i128 + last_da_gas_price as i128
+        - da_p_comp
+        - da_d_comp;
     assert_eq!(expected as u64, actual);
 }
 
@@ -120,7 +122,7 @@ fn calculate__price_does_not_decrease_more_than_max_percent() {
     let latest_gas_per_byte = 10;
     let da_p_component = 100;
     let da_d_component = 10;
-    let block_bytes = 500;
+    let block_bytes = 500u64;
     let last_profit = 100000; // Large, positive profit to decrease da price
     let last_last_profit = 0;
     let max_change_percent = 5;
@@ -132,7 +134,7 @@ fn calculate__price_does_not_decrease_more_than_max_percent() {
         .with_da_p_component(da_p_component)
         .with_da_d_component(da_d_component)
         .with_starting_da_gas_price(last_da_gas_price)
-        .with_total_rewards(larger_starting_reward)
+        .with_total_rewards(larger_starting_reward as u128)
         .with_known_total_cost(starting_cost as u128)
         .with_projected_total_cost(starting_cost as u128)
         .with_da_cost_per_byte(latest_gas_per_byte as u128)
@@ -159,7 +161,7 @@ fn calculate__da_price_does_not_increase_more_than_max_percent() {
     let latest_gas_per_byte = 10;
     let da_p_component = 1000;
     let da_d_component = 10;
-    let block_bytes = 500;
+    let block_bytes = 500u64;
     let last_profit = -1000000; // large, negative profit to increase da price
     let last_last_profit = 0;
     let arb_value = 1000;
@@ -172,7 +174,7 @@ fn calculate__da_price_does_not_increase_more_than_max_percent() {
         .with_starting_da_gas_price(last_da_gas_price)
         .with_da_p_component(da_p_component)
         .with_da_d_component(da_d_component)
-        .with_total_rewards(smaller_starting_reward)
+        .with_total_rewards(smaller_starting_reward as u128)
         .with_known_total_cost(starting_cost as u128)
         .with_projected_total_cost(starting_cost as u128)
         .with_da_cost_per_byte(latest_gas_per_byte as u128)
@@ -203,7 +205,7 @@ fn calculate__da_gas_price_never_drops_below_minimum() {
     let latest_gas_per_byte = 10;
     let da_p_component = 100;
     let da_d_component = 10;
-    let block_bytes = 500;
+    let block_bytes = 500u64;
     let profit_avg = 100;
     let avg_window = 10;
     let arb_value = 1000;
@@ -214,7 +216,7 @@ fn calculate__da_gas_price_never_drops_below_minimum() {
         .with_da_p_component(da_p_component)
         .with_da_d_component(da_d_component)
         .with_starting_da_gas_price(last_da_gas_price)
-        .with_total_rewards(larger_starting_reward)
+        .with_total_rewards(larger_starting_reward as u128)
         .with_known_total_cost(starting_cost as u128)
         .with_projected_total_cost(starting_cost as u128)
         .with_da_cost_per_byte(latest_gas_per_byte as u128)
